fixed function upvalue hydration bug.
It was being caused by objects used as upvalues being garbage collected in the middle of the
process of dumping values, and another object being created with the same memory address. since
the dump function maintains a map using memory address of objects as a key, the two independently-
created objects ended up with the same key.